Adapt source code to the new API

+++ src/memcache/memcache.cpp
@@ -58,7 +58,7 @@ inline UInt to_uint(const char* p, bool& result) {
         errno == ERANGE ) return 0;
     char c = *end;
     if( c != CR && c != LF && c != SP ) return 0;
-    if( i > std::numeric_limits<UInt>::max() ) return 0;
+    if( i > static_cast<unsigned long long>(std::numeric_limits<UInt>::max()) ) return 0;
     result = true;
     return static_cast<UInt>(i);
 }
